power: invert the monitor check
Previously, the semantics of power monitor were such that we were
checking current value against the expected value, and if they matched,
then the sleep was aborted. This is somewhat inflexible, because it only
allowed us to check for a specific value.

We can reverse the check, and instead have monitor sleep to be aborted
if the expected value *doesn't* match what's in memory. This allows us
to both implement all currently implemented driver code, as well as
support more use cases which don't easily map to previous semantics
(such as waiting on writes to AF_XDP counter value).

This commit also adjusts all current driver implementations to match the
new semantics.

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>