net/mlx5: fix loopback for DV queue
When switching to the Devx interface, the kernel driver behavior may
be different from using Verbs. The Tx loopback cannot work properly
even if the Tx and Rx queues are configured with loopback attribute.
To support self loopback for Tx, a Verbs dummy queue pair needs to
be created to trigger the kernel to enable the loopback.

This is only required when TIR is created for Rx and loopback is
needed. Only CQ and QP are needed for this case, no WQ(RQ) needs to
be created.

This requirement comes from bugzilla 312, more details can refer to:
https://bugs.dpdk.org/show_bug.cgi?id=645

Bugzilla ID: 645

Fixes: 6deb19e1b2d2 ("net/mlx5: separate Rx queue object creations")
Cc: stable@dpdk.org

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Acked-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>