net/mlx5: fix loopback for DV queue
In the past, all the queues and other hardware objects were created
through Verbs interface. Currently, most objects creation are
migrated to Devx interface by default, including queues. Only when
the DV is disabled by device arg or eswitch is enabled, are all or
some of the objects created through Verbs interface.

When using Devx interface to create queues, the kernel driver
behavior is different from using Verbs. The Tx loopback cannot work
properly even if the Tx and Rx queues are configured with loopback
attribute. To fix the support self loopback for Tx, a Verbs dummy
queue pair needs to be created to trigger the kernel to enable the
global loopback capability.

This is only required when TIR is created for Rx and loopback is
needed. Only CQ and QP are needed for this case, no WQ(RQ) needs to
be created.

This requirement comes from bugzilla 645, more details can be found
in the bugzilla link.

Bugzilla ID: 645

Fixes: 6deb19e1b2d2 ("net/mlx5: separate Rx queue object creations")
Cc: stable@dpdk.org

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Acked-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>