test: drop 'alarm_autotest' and 'cycles_autotest' from test suite
The tests 'alarm_autotest' and 'cycles_autotest' rely on the underlying
system having very accurate and precise timing.  On systems where the timing
isn't as rigid, or the load is particularly high, these tests are unreliable
since the wake latency from the scheduler can be high enough to miss the
timing window.

Remove these tests from the test suites.  Maybe it's useful for these
tests to be present as a diagnostics tool, but for normal unit testing,
they don't provide much value.  They have falsely flagged patches as
FAIL on various infrastructures.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>