fix Marvell maintainer email for atlantic
Fixing ex-Aquantia email - it is now part of Marvell.
Removing Pavel Belous email - he is not in company now.

Also adding Marvell company prefix to the driver names for
both atlantic and qede, to eliminate any confusion.

Qlogic is still actively used as a well known marketing name
for the device family, so keeping it.

Signed-off-by: Igor Russkikh <irusskikh@marvell.com>
Signed-off-by: Rasesh Mody <rmody@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>