net/vmxnet3: add interrupt support in MSI-X
Add support for MSI-x interrupt vectors to the vmxnet3 driver.
This will allow more efficient deployments in cloud environments.

By default it will try to allocate 1 vector (0) for link
event and one MSI-x vector for each rx queue.  To simplify
things, it will only be enabled if the number of tx and rx
queues are equal (so that tx/rx share the same vector).
If for any reason vmxnet3 cannot enable intr mode, it will
fall back to the lsc only mode.

Signed-off-by: Yong Wang <yongwang@vmware.com>
Signed-off-by: Jochen Behrens <jbehrens@vmware.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>