raw/ioat: extend python script functionality
Currently the user needs to find the DSA instance number for any DSA
device they would like to configure using this script, which can be
cumbersome and error-prone since the instance numbering changes when
changing the binding of the devices between vfio-pci and idxd. In
addition to this, once the device is configured, there is no option to
reset the device using the script.

This patch contains two additions to the script:
1. Add the ability to specify the DSA device to configure using the
   device's PCI address instead of the DSA instance number. For example,
   "$dpdk_idxd_cfg.py 0" and "$dpdk_idxd_cfg.py 6a:01.0" are both valid
   references to the same device (assuming the numbering).
2. An option to reset the device via the script. For example,
   "$dpdk_idxd_cfg.py 6a:01.0 --reset"

Signed-off-by: Kevin Laatz <kevin.laatz@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>